Compatibilty with drush v.4.0-[dev rc*] :
* Replaced all occurence of drush_op('system', $exec) with drush_op_system($exec)
because it is recommended to use drush_op_system().

* In a special case drush_sql_dump($db_spec) returns an array now,
the dump command is at the first index ($exec[0]),
so multi-sql-dump without --dump-type=full (default) failed with drush v. 4.0-dev,
becaues an array was passed drush_op_system().

* Added hook_drush_init() to check for drush v.4.0-[dev rc*]
with a builddate newer or equal 27. Dec 2010 as requirement
via checking DRUSH_VERSION and datestamp.

* Removed backup symlink within document root, becaue pm-update aborts,
if backup exists within the installation.

* Updated documentation accourding to the required version.